[Refactoring] Bot agent 분리 및 simple_agent 리팩토링 (#10)
* chore: bot 제거

* refactor: refactor: simple agent 모듈 구조 개선 및 책임 분리
