--- bot.env.example
@@ -1,7 +1,0 @@
-# bot_client용 (별도로 Docker로 띄울 예정)
-OPS_API_URL=<https://n.sodam.store>
-ADMIN_ACCESS_TOKEN=<재완님>
-# AWS Credentials (for Transcribe, Polly, Bedrock)
-AWS_ACCESS_KEY_ID=<상연님>
-AWS_SECRET_ACCESS_KEY=<상연님>
-AWS_DEFAULT_REGION=us-east-1
