--- voice/__init__.py
@@ -1,1 +1,0 @@
-"""Voice AI Agent Module"""
