--- voice/agents/__init__.py
@@ -1,4 +1,0 @@
-"""Voice agent implementations."""
-from .elderly_companion import ElderlyCompanionAgent
-
-__all__ = ["ElderlyCompanionAgent"]
