--- voice/tools/__init__.py
@@ -1,3 +1,0 @@
-"""Voice agent tools."""
-# Tools are defined as methods in Agent classes using @function_tool decorator
-# This module can be used for standalone tool utilities if needed
